@@ -1,33 +1,35 @@
 {
-	"name": "task_5",
-	"version": "1.0.0",
-	"description": "",
-	"main": "index.js",
-	"scripts": {
-		"test": "echo \"Error: no test specified\" && exit 1",
-		"build": "webpack build --config ./config/webpack.config.js",
-		"start": "webpack serve --config ./config/webpack.config.js"
-	},
-	"keywords": [],
-	"author": "",
-	"license": "ISC",
-	"devDependencies": {
-		"@babel/core": "^7.16.5",
-		"@babel/plugin-transform-runtime": "^7.16.7",
-		"@babel/preset-env": "^7.16.5",
-		"@babel/preset-react": "^7.16.5",
-		"babel-loader": "^8.2.3",
-		"css-loader": "^6.5.1",
-		"file-loader": "^6.2.0",
-		"html-webpack-plugin": "^5.5.0",
-		"image-webpack-loader": "^8.0.1",
-		"react-hot-loader": "^4.13.0",
-		"style-loader": "^3.3.1",
-		"webpack": "^5.65.0",
-		"webpack-cli": "^4.9.1",
-		"webpack-dev-server": "^4.6.0"
-	},
-	"dependencies": {
-		"@babel/runtime": "^7.16.7"
-	}
+  "name": "dashboard",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "jest",
+    "build": "webpack build --config ./config/webpack.config.js",
+    "start": "webpack-dev-server --mode development --config config/webpack.config.js"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "@babel/core": "^7.18.6",
+    "@babel/plugin-transform-runtime": "^7.18.6",
+    "@babel/preset-env": "^7.18.6",
+    "@babel/preset-react": "^7.18.6",
+    "babel-loader": "^8.2.5",
+    "css-loader": "^6.7.1",
+    "file-loader": "^6.2.0",
+    "html-webpack-plugin": "^5.5.0",
+    "image-webpack-loader": "^8.1.0",
+    "jest": "^28.1.3",
+    "style-loader": "^3.3.1",
+    "webpack": "^5.73.0",
+    "webpack-cli": "^4.10.0",
+    "webpack-dev-server": "^4.9.3"
+  },
+  "dependencies": {
+    "@babel/runtime": "^7.18.6",
+    "@zarconontol/enzyme-adapter-react-18": "^0.7.3",
+    "inline-source-map": "^0.6.2"
+  }
 }
