@@ -1,21 +1,17 @@
-import logo from './logo.svg';
-import './App.css';
+import logo from "./holberton-logo.jpg";
+import "./App.css";
 
 function App() {
   return (
     <div className="App">
-        <div className="App-header">
-        <img src={logo} alt="logo" />
+      <div className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
         <h1>School dashboard</h1>
       </div>
       <div className="App-body">
-        <p>
-          Login to access the full dashboard
-        </p>
+        <p>Login to access the full dashboard</p>
       </div>
-      <div className="App-footer">
-        <p>Copyright 2020 - holberton School</p>
-      </div>
+      <div className="App-footer">Copyright 2020 - holberton School</div>
     </div>
   );
 }
