@@ -1,13 +1,12 @@
+import React from "react";
+import "./Notifications.css";
 
-import React from 'react';
-import './Notifications.css';
+function Notifications() {
+  return (
+    <div className="Notifications">
+      <p>Here is the list of notifications</p>
+    </div>
+  );
+}
 
-export default function Notifications() {
-	return (
-		<div className="Notifications">
-			<p>
-				Here is the list of notifications
-			</p>
-		</div>
-	);
-};
+export default Notifications;
