@@ -1,27 +1,26 @@
-import logo from './logo.jpg';
-import './App.css';
-import { getFooterCopy, getFullYear } from './utils';
+import logo from "./holberton-logo.jpg";
+import "./App.css";
+import { getFullYear, getFooterCopy } from "./utils";
 
 function App() {
   return (
     <div className="App">
-        <div className="App-header">
-        <img src={logo} alt="logo" />
+      <div className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
         <h1>School dashboard</h1>
       </div>
       <div className="App-body">
-        <p>
-          Login to access the full dashboard
-        </p>
-		<br></br>
-		<label htmlFor="email">  Email: </label>
-        <input type="email" id="email" name="email" />
-        <label htmlFor="password">  Password: </label>
-        <input type="password" id="password" name="password" />
-        <button>  OK</button>
+        <p>Login to access the full dashboard</p>
+        <form>
+          <label htmlFor="email">Email:</label>
+          <input type="email" name="email"></input>
+          <label htmlFor="password">Password:</label>
+          <input type="password" name="password"></input>
+          <button>OK</button>
+        </form>
       </div>
       <div className="App-footer">
-	  <p>Copyright {getFullYear()} - {getFooterCopy()}</p>
+        Copyright {getFullYear()} - {getFooterCopy()}
       </div>
     </div>
   );
