@@ -1,4 +1,4 @@
-import { configure } from 'enzyme';
-import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
+import { configure } from "enzyme";
+import Adapter from "@zarconontol/enzyme-adapter-react-18";
 
 configure({ adapter: new Adapter() });
