@@ -1,30 +1,26 @@
-import React from 'react';
-import './Notifications.css';
-import closeIcon from '../assets/close-icon.png';
-import { getLatestNotification } from '../utils/utils';
+import React from "react";
+import "./Notifications.css";
+import closeIcon from "../assets/close-icon.png";
+import { getLatestNotification } from "../utils/utils";
 
-const Notifications = () => {
-	return (
-		<div className='Notifications'>
-			<p style={{ display: 'inline', marginRight: '80%' }}>
-				Here is the list of notifications
-			</p>
-			<button
-				aria-label='Close'
-				onClick={console.log('Close button has been clicked')}
-			>
-				<img style={{ display: 'inline' }} src={closeIcon} alt='Close' />
-			</button>
-			<ul>
-				<li data='default'>New course available</li>
-				<li data='urgent'>New resume available</li>
-				<li
-					data='urgent'
-					dangerouslySetInnerHTML={{ __html: getLatestNotification() }}
-				></li>
-			</ul>
-		</div>
-	);
-};
+function Notifications() {
+  return (
+    <div className="Notifications">
+      <button
+        style={{ color: "#3a3a3a", fontWeight: "bold", background: "none", border: "none", fontSize: "15px", position: "absolute", right: "2px", top: "2px", cursor: "pointer" }}
+        aria-label="Close"
+        onClick={console.log("Close button has been clicked")}
+      >
+        <img src={closeIcon} alt="closeIcon" width="10px" />
+      </button>
+      <p>Here is the list of notifications</p>
+      <ul>
+        <li data="default">New course available</li>
+        <li data="urgent">New resume available</li>
+        <li data="urgent" dangerouslySetInnerHTML={{ __html: getLatestNotification() }}></li>
+      </ul>
+    </div>
+  );
+}
 
 export default Notifications;
